bibliotheque/requirements.txt
chalice
pyjwt
